requirements.txt
openai==1.60.1
python-dotenv==1.0.1
Requests==2.32.3
streamlit==1.40.1
google-api-python-client
google-auth-httplib2 
google-auth-oauthlib
tzlocal
langchain